iox-#2457 Move path and file verifier functions to posix vocabulary, where they are used
