+++ iceoryx_hoofs/posix/vocabulary/source/file_path.cpp
@@ -15,11 +15,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "iox/file_path.hpp"
-#include "iox/filesystem.hpp"
+#include "iox/detail/path_and_file_verifier.hpp"
+#include "iox/string.hpp"
 
 namespace iox
 {
-namespace details
+namespace detail
 {
 bool file_path_does_contain_invalid_characters(const string<platform::IOX_MAX_PATH_LENGTH>& value) noexcept
 {
@@ -30,11 +31,11 @@ bool file_path_does_contain_invalid_characters(const string<platform::IOX_MAX_PA
         // AXIVION Next Construct AutosarC++19_03-A3.9.1: Not used as an integer but as actual character
         const char c{value.unchecked_at(i)};
 
-        const bool isSmallLetter{internal::ASCII_A <= c && c <= internal::ASCII_Z};
-        const bool isCapitalLetter{internal::ASCII_CAPITAL_A <= c && c <= internal::ASCII_CAPITAL_Z};
-        const bool isNumber{internal::ASCII_0 <= c && c <= internal::ASCII_9};
-        const bool isSpecialCharacter{c == internal::ASCII_DASH || c == internal::ASCII_DOT
-                                      || c == internal::ASCII_COLON || c == internal::ASCII_UNDERSCORE};
+        const bool isSmallLetter{detail::ASCII_A <= c && c <= detail::ASCII_Z};
+        const bool isCapitalLetter{detail::ASCII_CAPITAL_A <= c && c <= detail::ASCII_CAPITAL_Z};
+        const bool isNumber{detail::ASCII_0 <= c && c <= detail::ASCII_9};
+        const bool isSpecialCharacter{c == detail::ASCII_DASH || c == detail::ASCII_DOT || c == detail::ASCII_COLON
+                                      || c == detail::ASCII_UNDERSCORE};
 
         const bool isPathSeparator{[&] {
             for (const auto separator : platform::IOX_PATH_SEPARATORS)
@@ -60,5 +61,5 @@ bool file_path_does_contain_invalid_content(const string<platform::IOX_MAX_PATH_
 {
     return !isValidPathToFile(value);
 }
-} // namespace details
+} // namespace detail
 } // namespace iox
