+++ iceoryx_hoofs/test/moduletests/test_vocabulary_string_ctor_and_assign.cpp
@@ -171,12 +171,12 @@ TYPED_TEST(stringTyped_test, SelfMoveAssignmentExcluded)
 {
     ::testing::Test::RecordProperty("TEST_ID", "0ad45975-b68b-465a-b8c5-83dd8d8290d5");
     this->testSubject = "M";
-#if (defined(__GNUC__) && (__GNUC__ >= 13 && __GNUC__ <= 14) && !defined(__clang__))
+#if (defined(__GNUC__) && (__GNUC__ >= 13 && __GNUC__ <= 15) && !defined(__clang__))
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wself-move"
 #endif
     this->testSubject = std::move(this->testSubject);
-#if (defined(__GNUC__) && (__GNUC__ >= 13 && __GNUC__ <= 14) == 13 && !defined(__clang__))
+#if (defined(__GNUC__) && (__GNUC__ >= 13 && __GNUC__ <= 15) && !defined(__clang__))
 #pragma GCC diagnostic pop
 #endif
     EXPECT_THAT(this->testSubject.size(), Eq(1U));
